CatBonds class working
